+++ frontend/src/components/Menus/SideNavigationMenu.vue
